wasm

+++ src/CrisisConnect/CrisisConnect /Application/Services/Repositories/IOperationClaimRepository.cs
@@ -3,4 +3,4 @@ using Core.Security.Entities;
 
 namespace Application.Services.Repositories;
 
-public interface IOperationClaimRepository : IAsyncRepository<OperationClaim, long>, IRepository<OperationClaim, long> { }
+public interface IOperationClaimRepository : IAsyncRepository<OperationClaim, int>, IRepository<OperationClaim, int> { }
