+++ src/CrisisConnect/CrisisConnect /Application/Services/Repositories/IRefreshTokenRepository.cs
@@ -3,7 +3,7 @@ using Core.Security.Entities;
 
 namespace Application.Services.Repositories;
 
-public interface IRefreshTokenRepository : IAsyncRepository<RefreshToken, long>, IRepository<RefreshToken, long>
+public interface IRefreshTokenRepository : IAsyncRepository<RefreshToken, int>, IRepository<RefreshToken, int>
 {
     Task<List<RefreshToken>> GetOldRefreshTokensAsync(int userID, int refreshTokenTTL);
 }
