+++ src/CrisisConnect/CrisisConnect /Persistence/Repositories/RefreshTokenRepository.cs
@@ -6,7 +6,7 @@ using Persistence.Contexts;
 
 namespace Persistence.Repositories;
 
-public class RefreshTokenRepository : EfRepositoryBase<RefreshToken, long, BaseDbContext>, IRefreshTokenRepository
+public class RefreshTokenRepository : EfRepositoryBase<RefreshToken, int, BaseDbContext>, IRefreshTokenRepository
 {
     public RefreshTokenRepository(BaseDbContext context)
         : base(context) { }
