+++ src/CrisisConnect/CrisisConnect /Persistence/Repositories/UserOperationClaimRepository.cs
@@ -6,7 +6,7 @@ using Persistence.Contexts;
 
 namespace Persistence.Repositories;
 
-public class UserOperationClaimRepository : EfRepositoryBase<UserOperationClaim, long, BaseDbContext>, IUserOperationClaimRepository
+public class UserOperationClaimRepository : EfRepositoryBase<UserOperationClaim, int, BaseDbContext>, IUserOperationClaimRepository
 {
     public UserOperationClaimRepository(BaseDbContext context)
         : base(context) { }
